move provider to ntske package
